Follow youtube tutorila
I went through the whole tutorial. There were some syntax errors  that I made and hade to go looking through the code to see what I had done. I accidentally typed --Left instead of --left on one occasion. And the other, In the video he didn't put curly braces around a return value for a discrete function and i glossed over that and I put in the curly braces and I didn't see that he didn't do it that way so I had to add in a return. The lose mechanic wasn't working because of two little curly braces. I would curse if I wasn't so tired.